refactor(handlers): rename to _handlers
